Added heuristics and basic framework of pdhs.
Heuristic class and HansenHeuristic added to guide AO* search. Added simple foward search algorithm, but I think it still needs to pass down upper and lower bound functions to the Tree (and maybe to the Nodes as well).

Added GT, pseudoPOMDP, and POMDP models in for initial parsing and conversions.

Added basic framework of the policy driven heuristic search algorithm, but the main part is still missing.

Updated documentation.